requirements.txt
asgiref==3.5.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
Django==4.1
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.3.0
psycopg2-binary==2.9.8
pycparser==2.21
PySocks==1.7.1
pytz==2022.2.1
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.22.2
trio-websocket==0.11.1
tzdata==2022.2
urllib3==2.0.7
wsproto==1.2.0
